requirements.txt
Flask
Pillow
requests
spidev
gpiozero
